refactor: 右手側のファイル名をシンプルにする
