+++ src/button-accordion-with-keyboard/atoms/shared/types.ts
@@ -1,4 +1,4 @@
-import { ReedName } from "../../config/rightHand/rightHandConfig";
+import { ReedName } from "../../config/rightHand/config";
 
 // リードの有効/無効状態の型
 export type ReedActivation = Record<ReedName, boolean>;
