+++ src/button-accordion-with-keyboard/components/shared/KeyboardButton.tsx
