+++ src/button-accordion-with-keyboard/config/rightHand/config.ts
